% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cb_dlnm}
\alias{cb_dlnm}
\title{Cross basis function}
\usage{
cb_dlnm(db, var)
}
\arguments{
\item{db}{(default, NULL) is a ycc_lag database with the cases_4w
\itemize{
\item deaths_4w + cfr_4w + attack_4w variables
}}

\item{var}{(default, NULL) one of the variables in ycc_lag}
}
\value{
The crossbasis function
}
\description{
Creates a crossbasis function for a model to predict Status from a single lagged
epidemiological variable, with up to 4 weeks lag
}
\examples{
\dontrun{
  cb_dlnm(ycc_lag, "cases_4w")
}
}
