% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_log_epi}
\alias{get_log_epi}
\title{Region specific estimates of Log RR and errors}
\usage{
get_log_epi(db)
}
\arguments{
\item{db}{(default, NULL) is a ycc_lag database with the cases_4w
\itemize{
\item deaths_4w + cfr_4w + attack_4w variables
}}
}
\value{
A tibble with the estimates and standard errors for the
3 epidemiological variables (attack rate is excluded since it's
collinear by definition)
}
\description{
Get region specific estimates to be pooled later into a meta-
analysis through a glm model fitted on the usual 4 epidemiological
variables
}
\examples{
\dontrun{
  get_log_epi(ycc_lag)
}
}
