% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_dlnm}
\alias{model_dlnm}
\title{Lag-linear model with cross basis function}
\usage{
model_dlnm(db, cb_var)
}
\arguments{
\item{db}{(default, NULL) is a ycc_lag database with the cases_4w
\itemize{
\item deaths_4w + cfr_4w + attack_4w variables
}}

\item{cb_var}{(default, NULL) the crossbasis function}
}
\value{
The fitted model
}
\description{
Creates a model to predict Status from a single lagged
epidemiological variable, with up to 4 weeks lag
}
\examples{
\dontrun{
  model_dlnm(ycc_lag, "cases_4w")
}
}
