% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_spline}
\alias{model_spline}
\title{Time series spline model}
\usage{
model_spline(db)
}
\arguments{
\item{db}{(default, NULL) is a ycc_lag database with the cases_4w
\itemize{
\item deaths_4w + cfr_4w + attack_4w variables
}}
}
\value{
confidence interval after fitting a GLM with 10 df spline
to model temporal dynamics
}
\description{
Creates a glm model to predict Status from other epidemiological
variables
}
\examples{
\dontrun{
  model_spline(ycc_lag)
}
}
