% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_ycc}
\alias{preprocess_ycc}
\title{Preprocessing of Yemen Cholera cases}
\usage{
preprocess_ycc(db)
}
\arguments{
\item{db}{: dataframe obtained by the Yemen Cholera cases csv}
}
\value{
a new db with
\itemize{
\item clean column names,
\item dates subdivided by epidemiological week and year,
\item with all epi variables grouped by week,
\item government names that are the same as OCHA adm boundaries
}
}
\description{
Preprocessing of Yemen Cholera cases
}
\examples{
\dontrun{
  preprocess_ycc(ycc)
}
}
