% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_ycc_lags}
\alias{preprocess_ycc_lags}
\title{Preprocessing of Yemen Cholera cases}
\usage{
preprocess_ycc_lags(db)
}
\arguments{
\item{db}{: Yemen Cholera Cases (ycc) database}
}
\value{
a new db with new cases, new deaths, CFR, Attack rate
within the last 4 weeks
}
\description{
Preprocessing of Yemen Cholera cases
}
\examples{
\dontrun{
  preprocess_ycc_lags(ycc)
}
}
