% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ycc_exp}
\alias{ycc_exp}
\title{Calculate Expected counts}
\usage{
ycc_exp(db)
}
\arguments{
\item{db}{: a ycc_lag database}
}
\value{
db: ycc_lag with additional columns for expected epi
variables and lead status (target of prediction).
To provide better estimates the cases are averaged over the last 4
weeks (hence the use of ycc_lag)
}
\description{
Calculate Expected counts
}
\examples{
#' \dontrun{
  ycc_exp(ycc_lag)
}
}
