% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeraGeo4H-package.R
\docType{package}
\name{CholeraGeo4H-package}
\alias{CholeraGeo4H}
\alias{CholeraGeo4H-package}
\title{CholeraGeo4H: Yemen Cholera Outbreak Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Geospatial analysis of the 2017-2018 Yemen Cholera Outbreak dynamics.
}
\author{
\strong{Maintainer}: Andrea Pedot \email{andrea.pedot@ubep.unipd.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
